pyproject.toml
[project]
name = "ai-pulse"
version = "0.1.0"
description = "Automated daily AI briefing system"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "feedparser>=6.0.11",
    "schedule>=1.2.2",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.1",
    "jinja2>=3.1.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pipeline", "dashboard"]

[tool.uv]
dev-dependencies = []
